% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache}
\alias{cache}
\alias{cache_dir}
\alias{cache_info}
\alias{cache_delete}
\title{Work with the cache directory}
\usage{
cache_dir()

cache_info(cache_dir = rappdirs::user_cache_dir("tiltWorkflows"))

cache_delete(
  cache_dir = rappdirs::user_cache_dir("tiltWorkflows"),
  quiet = FALSE
)
}
\arguments{
\item{cache_dir}{Character. A directory.}

\item{quiet}{Logical. Throw a warning?}
}
\value{
\itemize{
\item \code{cache_dir()} returns a path to the default cache directory.
\item \code{cache_info()} returns a data frame with information about the cache directory.
\item \code{cache_delete()} deletes the cache directory.
}
}
\description{
Work with the cache directory
}
\examples{
library(withr)
library(fs)

cache_dir()

tmp <- dir_create(local_tempfile())
cache_info(tmp)

dir_exists(tmp)
cache_delete(tmp)
dir_exists(tmp)
}
