% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_demo.R
\name{cleanup}
\alias{cleanup}
\alias{cleanup_io}
\alias{cleanup_profile_emissions}
\title{Cleanup artifacts of a demo workflow}
\usage{
cleanup_io()

cleanup_profile_emissions()
}
\value{
Called for its side effects of deleting files and folders.
}
\description{
Cleanup artifacts of a demo workflow
}
\examples{
cleanup_io()
cleanup_profile_emissions()
}
\keyword{internal}
