% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{explain}
\alias{explain}
\alias{explain_indicator}
\alias{explain_results}
\title{Explain stuff consistently everywhere}
\usage{
explain_indicator()

explain_results()
}
\value{
Character.
}
\description{
Explain stuff consistently everywhere
}
\examples{
explain_indicator()

explain_results()
}
\keyword{internal}
