% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{explain_indicator}
\alias{explain_indicator}
\alias{explain_resutls}
\title{Explain sections consistently accross workflows.}
\usage{
explain_indicator()

explain_resutls()
}
\value{
Character.
}
\description{
Explain sections consistently accross workflows.
}
\examples{
explain_indicator()
}
\keyword{internal}
