% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIXME_issue_49.R
\name{FIXME_issue_49}
\alias{FIXME_issue_49}
\alias{FIXME_issue_49_ecoinvent_inputs}
\title{See issue #49}
\usage{
FIXME_issue_49(data)

FIXME_issue_49_ecoinvent_inputs(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A data frame
}
\description{
See issue #49
}
\examples{
#
}
\keyword{internal}
