% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile_sector}
\alias{profile_sector}
\title{Profile emissions and upstream emissions}
\usage{
profile_sector(
  companies,
  scenarios,
  europages_companies,
  ecoinvent_activities,
  ecoinvent_europages,
  isic_tilt,
  low_threshold = ifelse(scenarios$year == 2030, 1/9, 1/3),
  high_threshold = ifelse(scenarios$year == 2030, 2/9, 2/3)
)
}
\arguments{
\item{europages_companies}{Dataframe. Companies from europages.}

\item{ecoinvent_activities}{Dataframe. Activities from ecoinvent.}

\item{ecoinvent_europages}{Dataframe. Mapper between europages and ecoinvent.}

\item{isic_tilt}{Dataframe. Mapper between isic and tilt.}

\item{low_threshold}{A numeric value to segment low and medium transition
risk products.}

\item{high_threshold}{A numeric value to segment medium and high transition
risk products.}
}
\value{
A data frame with the column \code{companies_id}, and the nested columns\code{product} and \code{company} holding the outputs at product and company level. Unnesting \code{product} yields a data frame with at least columns \code{companies_id}, \code{grouped_by}, \code{risk_category}. Unnesting \code{company} yields a data frame with at least columns \code{companies_id}, \code{grouped_by}, \code{risk_category}, \code{value}.
}
\description{
These functions wrap the output of the corresponding function in
\href{https://2degreesinvesting.github.io/tiltIndicator/reference/index.html}{tiltIndicator}.
}
\examples{
library(tiltToyData)
library(readr, warn.conflicts = FALSE)

options(readr.show_col_types = FALSE)

companies <- read_csv(toy_emissions_profile_any_companies())
products <- read_csv(toy_emissions_profile_products())

result <- profile_emissions(
  companies,
  products,
  # TODO: Move to tiltToyData
  europages_companies = tiltIndicatorAfter::ep_companies,
  ecoinvent_activities = tiltIndicatorAfter::ecoinvent_activities,
  ecoinvent_europages = tiltIndicatorAfter::matches_mapper |> head(100),
  isic_tilt = tiltIndicatorAfter::isic_tilt_mapper
)

result |> unnest_product()

result |> unnest_company()



inputs <- read_csv(toy_emissions_profile_upstream_products())

result <- profile_emissions_upstream(
  companies,
  inputs,
  # TODO: Move to tiltToyData
  europages_companies = tiltIndicatorAfter::ep_companies,
  ecoinvent_activities = tiltIndicatorAfter::ecoinvent_activities,
  ecoinvent_inputs = tiltIndicatorAfter::ecoinvent_inputs,
  ecoinvent_europages = tiltIndicatorAfter::matches_mapper |> head(100),
  isic_tilt = tiltIndicatorAfter::isic_tilt_mapper
)

result |> unnest_product()

result |> unnest_company()
}
\seealso{
Other top-level functions: 
\code{\link[tiltIndicatorAfter]{profile_sector}()}
}
\keyword{internal}
