% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile_sector_upstream}
\alias{profile_sector_upstream}
\title{Profile sector and upstream sector}
\usage{
profile_sector_upstream(
  companies,
  scenarios,
  inputs,
  europages_companies,
  ecoinvent_activities,
  ecoinvent_inputs,
  ecoinvent_europages,
  isic_tilt,
  low_threshold = ifelse(scenarios$year == 2030, 1/9, 1/3),
  high_threshold = ifelse(scenarios$year == 2030, 2/9, 2/3)
)
}
\arguments{
\item{companies, scenarios, inputs}{A dataframe like the dataset with a matching name in tiltToyData (see \href{https://2degreesinvesting.github.io/tiltToyData/reference/index.html}{Reference}).}

\item{europages_companies}{Dataframe. Companies from europages.}

\item{ecoinvent_activities}{Dataframe. Activities from ecoinvent.}

\item{ecoinvent_inputs}{Dataframe. Upstream products from ecoinvent.}

\item{ecoinvent_europages}{Dataframe. Mapper between europages and ecoinvent.}

\item{isic_tilt}{Dataframe. Mapper between isic and tilt.}

\item{low_threshold}{A numeric value to segment low and medium transition
risk products.}

\item{high_threshold}{A numeric value to segment medium and high transition
risk products.}
}
\value{
A data frame with the column \code{companies_id}, and the nested columns\code{product} and \code{company} holding the outputs at product and company level. Unnesting \code{product} yields a data frame with at least columns \code{companies_id}, \code{grouped_by}, \code{risk_category}. Unnesting \code{company} yields a data frame with at least columns \code{companies_id}, \code{grouped_by}, \code{risk_category}, \code{value}.
}
\description{
These functions wrap the output of the corresponding function in
\href{https://2degreesinvesting.github.io/tiltIndicator/reference/index.html}{tiltIndicator}.
}
\examples{
library(tiltToyData)
library(readr, warn.conflicts = FALSE)

options(readr.show_col_types = FALSE)

companies <- read_csv(toy_sector_profile_companies())
scenarios <- read_csv(toy_sector_profile_any_scenarios())

result <- profile_sector(
  companies,
  scenarios,
  # TODO: Move to tiltToyData
  europages_companies = ep_companies |> head(3),
  ecoinvent_activities = ecoinvent_activities |> head(3),
  ecoinvent_europages = matches_mapper |> head(3),
  isic_tilt = isic_tilt_mapper |> head(3)
)

result |> unnest_product()

result |> unnest_company()



companies <- read_csv(toy_sector_profile_upstream_companies())
scenarios <- read_csv(toy_sector_profile_any_scenarios())
inputs <- read_csv(toy_sector_profile_upstream_products())
ecoinvent_inputs <- ecoinvent_inputs |> head(3)

result <- profile_sector_upstream(
  companies,
  scenarios,
  inputs,
  # TODO: Move to tiltToyData
  europages_companies = ep_companies |> head(3),
  ecoinvent_activities = ecoinvent_activities |> head(3),
  ecoinvent_inputs = ecoinvent_inputs |> head(3),
  ecoinvent_europages = matches_mapper |> head(3),
  isic_tilt = isic_tilt_mapper |> head(3)
)

result |> unnest_product()

result |> unnest_company()
}
\seealso{
Other top-level functions: 
\code{\link[tiltIndicatorAfter]{profile_emissions}()}
}
\keyword{internal}
