% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-profile.R
\name{profile_with}
\alias{profile_with}
\title{Run a \verb{profile*()} function by chunks}
\usage{
profile_with(...)
}
\arguments{
\item{...}{Arguments passed to \code{.f}.}

\item{companies}{Data frame.}

\item{.f}{A \verb{profile*()} function from tiltIndicatorAfter.}

\item{.by}{Character. Columns to chunk by.}

\item{chunks}{Integer. Number of chunks to split the data by.}
}
\value{
A nested data frame with results at product and company level.
}
\description{
Run a \verb{profile*()} function by chunks
}
\examples{
library(tiltToyData)
library(readr, warn.conflicts = FALSE)

options(readr.show_col_types = FALSE)

companies <- read_csv(toy_emissions_profile_any_companies())
products <- read_csv(toy_emissions_profile_products())

result <- companies |>
  profile_with(profile_emissions,
    products,
    # TODO: Move to tiltToyData
    europages_companies = tiltIndicatorAfter::ep_companies,
    ecoinvent_activities = tiltIndicatorAfter::ecoinvent_activities,
    ecoinvent_europages = tiltIndicatorAfter::matches_mapper |> head(100),
    isic_tilt = tiltIndicatorAfter::isic_tilt_mapper
  )

result |> unnest_product()

result |> unnest_company()
}
\keyword{internal}
