% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_falling_back_to.R
\name{read_csv_falling_back_to}
\alias{read_csv_falling_back_to}
\title{Read a .csv file; if it doesn't exist, fall back to another dataset}
\usage{
read_csv_falling_back_to(path, data)
}
\arguments{
\item{path}{Character. A path to a .csv file.}

\item{data}{Data frame. A dataset to fall back to if the .csv file doesn't
exist.}
}
\value{
A data frame.
}
\description{
Read a .csv file; if it doesn't exist, fall back to another dataset
}
\examples{
read_csv_falling_back_to("/does/not/exist.csv", head(cars))
}
\keyword{internal}
