% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{score_transition_risk}
\alias{score_transition_risk}
\title{Transition Risk Score}
\usage{
score_transition_risk(
  emissions_profile_at_product_level,
  sector_profile_at_product_level
)
}
\arguments{
\item{emissions_profile_at_product_level}{Dataframe. Emissions profile product level output}

\item{sector_profile_at_product_level}{Dataframe. Sector profile product level output}
}
\value{
A dataframe
}
\description{
Calulate Transition Risk Score at product level and company level
}
\examples{
library(dplyr)
library(readr, warn.conflicts = FALSE)
library(tiltToyData)
options(readr.show_col_types = FALSE)

emissions_companies <- read_csv(toy_emissions_profile_any_companies())
products <- read_csv(toy_emissions_profile_products())
europages_companies <- read_csv(toy_europages_companies())
ecoinvent_activities <- read_csv(toy_ecoinvent_activities())
ecoinvent_europages <- read_csv(toy_ecoinvent_europages())
isic_name <- read_csv(toy_isic_name())

emissions_profile_at_product_level <- profile_emissions(
  companies = emissions_companies,
  co2 = products,
  europages_companies = europages_companies,
  ecoinvent_activities = ecoinvent_activities,
  ecoinvent_europages = ecoinvent_europages,
  isic = isic_name
) |> unnest_product()

sector_companies <- read_csv(toy_sector_profile_companies())
scenarios <- read_csv(toy_sector_profile_any_scenarios())

sector_profile_at_product_level <- profile_sector(
  companies = sector_companies,
  scenarios = scenarios,
  europages_companies = europages_companies,
  ecoinvent_activities = ecoinvent_activities,
  ecoinvent_europages = ecoinvent_europages,
  isic = isic_name
) |> unnest_product()

result <- score_transition_risk(emissions_profile_at_product_level, sector_profile_at_product_level)

result |> unnest_product()

result |> unnest_company()
}
\seealso{
Other top-level functions: 
\code{\link[tiltIndicatorAfter]{profile_emissions}()},
\code{\link[tiltIndicatorAfter]{profile_sector}()}
}
\keyword{internal}
