% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIXME_issue_49.R
\name{FIXME_issue_49}
\alias{FIXME_issue_49}
\title{See issue #49}
\usage{
FIXME_issue_49(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A data frame
}
\description{
See issue #49
}
\examples{
tibble::tibble(
  TO_REMOVE = "a",
  company_name = "a",
  country = "a",
  company_city = "a",
  postcode = "a",
  address = "a",
  main_activity = "a",
  companies_id = "a"
) |>
FIXME_issue_49()
}
\keyword{internal}
