% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{read_csv_falling_back_to}
\alias{read_csv_falling_back_to}
\title{Read a .csv but if it doesnt' exist fall back to another dataset}
\usage{
read_csv_falling_back_to(path, data)
}
\arguments{
\item{path}{Character. Path to a .csv file.}

\item{data}{Data frame. An alternative dataset to fall back to if the \code{path}
points to a file that doesn't exist.}
}
\value{
A data frame.
}
\description{
Read a .csv but if it doesnt' exist fall back to another dataset
}
\examples{
read_csv_falling_back_to("/inexistent/file.csv", head(cars))
}
