% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_toy_input.R
\name{use_toy_input}
\alias{use_toy_input}
\title{Write an input/ directory with toy datasets into the working directory}
\usage{
use_toy_input(quiet = FALSE)
}
\arguments{
\item{quiet}{Logical. Throw a message?}
}
\value{
Call for it's side effect.
}
\description{
Write an input/ directory with toy datasets into the working directory
}
\examples{
use_toy_input()

# It won't overwrite
try(use_toy_input())

list.files("input")

# Cleanup
unlink("input")
}
