% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_workflow.R
\name{use_workflow}
\alias{use_workflow}
\title{Use a workflow file}
\usage{
use_workflow(template = "profile.Rmd", save_as = template, open = FALSE)
}
\arguments{
\item{template}{Path to template file relative to \verb{templates/} directory
within \code{package}; see details.}

\item{save_as}{Path of file to create, relative to root of active project.
Defaults to \code{template}}

\item{open}{Open the newly created file for editing? Happens in RStudio, if
applicable, or via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise.}
}
\value{
A logical vector indicating if file was modified.
}
\description{
Creates a workflow file from a template in the tiltWorkflows package.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
# Note: Running this will write "profile.Rmd" to your working directory
use_workflow()
}
\dontshow{\}) # examplesIf}
}
