% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_workflow.R
\name{use_workflow}
\alias{use_workflow}
\alias{workflows}
\title{List and use workflows}
\usage{
use_workflow(
  template = "profile.Rmd",
  save_as = template,
  open = rlang::is_interactive()
)

workflows()
}
\arguments{
\item{template}{Path to template file relative to \verb{templates/} directory
within \code{package}; see details.}

\item{save_as}{Path of file to create, relative to root of active project.
Defaults to \code{template}}

\item{open}{Open the newly created file for editing? Happens in RStudio, if
applicable, or via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise.}
}
\value{
A logical vector indicating if file was modified.
}
\description{
\itemize{
\item \code{workflows()} lists available workflows.
\item \code{use_workflow()} creates a workflow file from a template in the tiltWorkflows
package.
}
}
\examples{
workflows()

\dontrun{
# Note: Running this will write "profile_emissions.Rmd" to your working directory
use_workflow("profile_emissions.Rmd")
}
}
